Improve error analysis output
The error analysis is restricted to a single core to make sure that
the logging output (bounds values) are included in the Sphinx documentation.
It is not clear what is causing the problem here (it may not be a Sherpa
issue).

Tweak the region-projection example, and include a manually created plot
using the data - image + contours.