Retrieve the manager and model data paths for XSPEC.
These two functions are not exported by the xspec module itself, only
the compiled version. This will be changed in a future commit.

It is not clear whether the tests are adequate, since the default paths
may be different for different builds, and can also be different if
environment variables are set (at least, the model data path can
be set by the XSPEC_MDATA_DIR environment variable, so that should be
checked by the test when creating the default value).