Upgrade test to support XSPEC_MDATA_DIR is set.
The model data path is taken from the XSPEC_MDATA_DIR environment variable
by the XSPEC initialization code, if set. Support this in the test, but
this support has not bene tested.

Is it worth testing the actual value, or just that it's set (true for
the model and manager paths)?